and more cleanup
